% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatar_data.R
\name{formatar_data}
\alias{formatar_data}
\title{Formatar datas para envio}
\usage{
formatar_data(data, modo = c("inicio", "fim"))
}
\arguments{
\item{data}{data incompleta}
}
\value{
data completa
}
\description{
Função formata datas incompletas para serem lidas pelo web service do BC
}
\keyword{internal}
