% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUltimoValorXML.R
\name{getUltimoValorXML}
\alias{getUltimoValorXML}
\title{Obter último valor de série de tempo em formato XML}
\usage{
getUltimoValorXML(serie)
}
\arguments{
\item{serie}{código da série no Sistema Gerenciador de Séries Temporais}
}
\value{
texto XML com resposta à consulta
}
\description{
Função obtém o último valor da série de tempo requisitada, utilizando método
getUltimoValorXML do web service do Banco Central. Para resultados mais
compreensíveis utilize a função \code{obter_bc}.
}
\examples{
\dontrun{dolar <- getUltimoValorXML(1)}
}
\seealso{
\code{\link{obter_bcws}}
\code{\link{obter_bc}}
}
