% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValor.R
\name{getValor}
\alias{getValor}
\title{Obter valor de série de tempo}
\usage{
getValor(serie, inicio = "01/01/1995")
}
\arguments{
\item{serie}{código da série no Sistema Gerenciador de Séries Temporais}

\item{inicio}{data de inicío da consulta}
}
\value{
valor da séria requisitada
}
\description{
Função obtém valor da série de tempo no período requisitado, utilizando método
getValor do web service do Banco Central.
}
\examples{
dolar <- getValor(1, "01/03/2012") #Obtém o valor de câmbio vigente em 01/03/2012
}
\seealso{
\code{\link{obter_bcws}}
\code{\link{obter_bc}}
}
