% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValorEspecial.R
\name{getValorEspecial}
\alias{getValorEspecial}
\title{Obter valores série de tempo especial}
\usage{
getValorEspecial(serie, inicio = "01/01/1995", fim = Sys.Date())
}
\arguments{
\item{serie}{código da série no Sistema Gerenciador de Séries Temporais}

\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}
}
\value{
valor da série especial requisitada
}
\description{
Função obtém valores da série de tempo especial requisitada, utilizando método
getValorEspecial do web service do Banco Central.
}
\seealso{
\code{\link{obter_bcws}}
\code{\link{obter_bc}}
}
