% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValoresSeriesXML.R
\name{getValoresSeriesXML}
\alias{getValoresSeriesXML}
\title{Obter série de tempo em formato XML}
\usage{
getValoresSeriesXML(serie, inicio = "01/01/1995", fim = Sys.Date())
}
\arguments{
\item{serie}{código da série no Sistema Gerenciador de Séries Temporais}

\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}
}
\value{
texto XML com resposta à consulta
}
\description{
Função obtém valores da série de tempo requisitada, utilizando método
getValoresSeriesXML do web service do Banco Central. Para resultados mais
compreensíveis utilize a função \code{obter_bc}.
}
\examples{
# Obtém toda a série de tempo de câmbio, a partir de 1995, em formato XML
dolar <- getValoresSeriesXML(1) 
# Obter a série de tempo de câmbio, a partir de 1995, em formato XML
dolar <- getValoresSeriesXML(1, inicio="01/01/1995", fim = "31/01/1995")
}
\seealso{
\code{\link{obter_bcws}}
\code{\link{obter_bc}}
}
