% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_bcws.R
\name{obter_bcws}
\alias{obter_bcws}
\title{Obter arquivo resposta para consulta on web service do Banco Central}
\usage{
obter_bcws(metodo = c("getValoresSeriesXML", "getUltimoValorXML", "getValor",
  "getValorEspecial"), serie, inicio = "01/01/1995", fim = Sys.Date())
}
\arguments{
\item{metodo}{método a ser realizado na consulta}

\item{serie}{código da série no Sistemas Gerenciador de Séries Temporais}

\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}
}
\value{
texto XML com resposta a consulta
}
\description{
Função que se encarrega de fazer a comunicação com o Web Service do Banco
Central
}
\seealso{
\code{\link{obter_bc}}
\code{\link{getValoresSeriesXML}}
\code{\link{getUltimoValorXML}}
\code{\link{getValor}}
\code{\link{getValorEspecial}}
}
\keyword{internal}
