% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_ipca.R
\name{obter_ipca}
\alias{obter_ipca}
\title{Realiza consulta de série de tempo do IPCA}
\usage{
obter_ipca(inicio = "01/01/1995", fim = Sys.Date(), tipo = c("indice",
  "var"), frequencia = c("mensal", "anual"), base = 1993)
}
\arguments{
\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}

\item{tipo}{tipo de variavel ("indice", "var")}

\item{frequencia}{frequencia da variavel ("mensal", "anual")}

\item{base}{ano base do número indice. Somente útil quando tipo = "indice". Padrão = 1993.}
}
\value{
Série de inflação medida pelo IPCA.
}
\description{
Função consulta a série de IPCA no webservice do BC.
}
\examples{
# Obter série de ipca, em número índice, a partir de 01/01/1995
dolar <- obter_ipca()
# Obter série de ipca, em variação percentual, a partir de 01/01/1995
dolar <- obter_ipca(tipo = "var")
# Obter série anual de ipca entre 2010 e 2017
dolar <- obter_ipca(inicio = 2010, fim = 2017, frequencia = "anual")
}
\seealso{
\code{\link{obter_bcws}}
\code{\link{getValoresSeriesXML}}
}
