% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_pib.R
\name{obter_pib}
\alias{obter_pib}
\title{Realiza consulta de série de tempo do PIB}
\usage{
obter_pib(inicio = "01/01/1995", fim = Sys.Date(), tipo = c("encad",
  "encaddess", "mensal", "mensalacum", "corrente", "percapita", "IBGE"))
}
\arguments{
\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}

\item{tipo}{tipo de variavel}
}
\value{
Série de tempo desejada.
}
\description{
Função consulta as séries de tempo do PIB no web service do BC
}
\seealso{
\code{\link{obter_bcws}}
\code{\link{getValoresSeriesXML}}
}
