% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_pib.R
\name{obter_pibpercapita}
\alias{obter_pibpercapita}
\title{Realiza consulta de série de tempo do PIB per capita}
\usage{
obter_pibpercapita(inicio = "01/01/1995", fim = Sys.Date())
}
\arguments{
\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}
}
\value{
Série de tempo do PIB per capita
}
\description{
Função é um cápsula em torno de \code{obter_pib}, que consulta a série de
PIB per capita.
}
\examples{
# Obter série de pib, em valores per capita, entre 2010 e 2014
pib <- obter_pibpercapita(inicio = 2010, fim = 2014)
}
\seealso{
\code{\link{obter_pib}}
}
