% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_sm.R
\name{obter_sm}
\alias{obter_sm}
\title{Realiza consulta de série de tempo do Salário Mínimo}
\usage{
obter_sm(inicio = "01/01/1995", fim = Sys.Date(), unidade = c("nominal",
  "real"), frequencia = c("mensal", "anual"), base = 1995)
}
\arguments{
\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}

\item{unidade}{define se valores reais ou nominais}

\item{frequencia}{frequencia da variavel}

\item{base}{ano base para valores reais}
}
\value{
Série de tempo de salário mínimo
}
\description{
Função consulta a série de tempo se salário mínimo no web service do BC.
}
\examples{
# Obter série de salário mínimo a partir de 01/01/1995
sm <- obter_sm()
# Obter série de salário mínimo em valores resais a partir de 01/01/1995
sm <- obter_sm(unidade = "real")
# Obter série anual de salario mínimo  entre 2010 e 2014
sm <- obter_sm(inicio = 2010, fim = 2014, frequencia = "anual")
}
\seealso{
\code{\link{obter_bcws}}
\code{\link{getValoresSeriesXML}}
}
