% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_request_xml.R
\name{soap_request_xml}
\alias{soap_request_xml}
\title{Construir arquivo de requisição SOAP}
\usage{
soap_request_xml(metodo, serie, inicio, fim)
}
\arguments{
\item{metodo}{Nome do metodo}

\item{serie}{código da série no Sistemas Gerados de Séries Temporais}

\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}
}
\value{
texto XML para requisição SOAP
}
\description{
Função cria texto XML para requisição SOAP
}
\seealso{
\code{\link{obter_bcws}}
}
\keyword{internal}
