% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_bc.R
\name{obter_bc}
\alias{obter_bc}
\title{Realiza consulta de série de tempo no sistema gerador de séries temporais do BC}
\usage{
obter_bc(codigos, inicio = "01/01/1995", fim = Sys.Date(),
  bloqueado.bool = FALSE, remove.old = TRUE)
}
\arguments{
\item{codigos}{codigos das séries de tempo desejadas}

\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}

\item{bloqueado.bool}{Boolean define se o campo "bloqueado" retornado pelo
web service do BC deve ser exibido ou não, padrão é FALSE.}

\item{remove.old}{Boolen define se os valores sem tratamento também devem
retornados, útil para testes. Padrão é FALSE.}
}
\value{
Série de tempo desejada.
}
\description{
Função genérica consulta a série de tempo desejada no web service do BC e a
retorna como uma série do tempo tradicional do R.
}
\examples{
# Obter série de câmbio a partir de 1995
obter_bc(1)
# Obter série de câmbio entre 01/01/1995 e 31/12/1995
obter_bc(1, inicio = "01/01/1995", fim = "31/12/1995")
# Mesmo que o example anterior e mais simples
obter_bc(1, inicio = 1995, fim = 1995)
# Obter série de compra e venda de dolár ente 01/01/1995 e 31/12/1995
obter_bc(c(1, 10813), inicio = 1995, fim = 1995)
}
\seealso{
\code{\link{obter_bcws}}
\code{\link{getValoresSeriesXML}}
}
