% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_pib.R
\name{obter_pibcorrente}
\alias{obter_pibcorrente}
\title{Realiza consulta de série de tempo do PIB em preços correntes}
\usage{
obter_pibcorrente(inicio = "01/01/1995", fim = Sys.Date())
}
\arguments{
\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}
}
\value{
Série de tempo do PIB corrente
}
\description{
Função é um cápsula em torno de \code{obter_pib}, que consulta a série de
PIB corrente anual.
}
\seealso{
\code{\link{obter_pib}}
}
