% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obter_pib.R
\name{obter_pibencad}
\alias{obter_pibencad}
\title{Realiza consulta de série de tempo do PIB encadeado}
\usage{
obter_pibencad(inicio = "01/01/1995", fim = Sys.Date())
}
\arguments{
\item{inicio}{data de inicío da consulta}

\item{fim}{data final da consulta}
}
\value{
Série de tempo do PIB encadeado.
}
\description{
Função é um cápsula em torno de \code{obter_pib}, que consulta a série de
PIB em número índice.
}
\examples{
# Obter série de pib, em valores encadeados, entre 2010 e 2014
pib <- obter_pibencad(inicio = 2010, fim = 2014)
}
\seealso{
\code{\link{obter_pib}}
}
